Pickup Progress
pickup script, health pickup, player manager

+++ CS 345 Final/Assets/Scripts/Health/HealthPickup.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class HealthPickup : MonoBehaviour
 {
-    [SerializeField] private float healthValue;
+    [SerializeField] private int healthValue;
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
